Fix bug that was causing a lose of events that were created in the same millisecond
