+++ packages/framework-provider-aws/test/library/subscription-adapter.test.ts
@@ -12,7 +12,7 @@ import {
   subscribeToReadModel,
   SubscriptionIndexRecord,
 } from '../../src/library/subscription-adapter'
-import { sortKeyForSubscription } from '../../src/library/partition-keys'
+import { sortKeyForSubscription } from '../../src/library/keys-helper'
 import { DocumentClient } from 'aws-sdk/lib/dynamodb/document_client'
 
 const logger: Logger = {
